Added queue and visual progress bar

+++ server/woocommerce/dev/dev.js
@@ -3,6 +3,8 @@ import { WooCommerce_Shops } from "../../config/woocommerce.js";
 import { Product } from "../../models/product.js";
 import fs from "fs";
 import { WooCreateProduct, WooUpdateQuantityProducts } from "../products.js";
+import PQueue from "p-queue";
+import { MultiBar, Presets } from 'cli-progress';
 
 async function findAndDeleteHidden() {
     // This function looks for products that are marked as hidden in our database and checks if they exist in WooCommerce.
@@ -62,11 +64,15 @@ async function syncWoo() {
     }
     console.log("Starting WooCommerce sync...");
 
-    for (let i = 0; i < WooCommerce_Shops.length; i++) {
-        const shop = WooCommerce_Shops[i];
-        console.log(`----------------------------------------`);
-        console.log(`Syncing products for shop: ${shop.url}`);
+    const multibar = new MultiBar({
+        clearOnComplete: false,
+        hideCursor: true,
+        format: ' {bar} | {percentage}% | {value}/{total} Products to {action} | Shop: {shop}'
+    }, Presets.shades_classic);
 
+    const tasks = [];
+
+    async function syncShopProducts(shop, i, deleteBar, updateBar, createBar) {
         const productsToDeletePath = `server/woocommerce/dev/results/productsToDelete_shop_${i}.json`;
         const productsQtyToUpdatePath = `server/woocommerce/dev/results/productsQtyToUpdate_shop_${i}.json`;
         const productsToCreatePath = `server/woocommerce/dev/results/productsToCreate_shop_${i}.json`;
@@ -76,37 +82,47 @@ async function syncWoo() {
         const productsToCreate = fs.existsSync(productsToCreatePath) ? JSON.parse(fs.readFileSync(productsToCreatePath)) : [];
 
         if (productsToDelete.length > 0) {
-            console.log(`Deleting ${productsToDelete.length} products in WooCommerce...`);
+            deleteBar.setTotal(productsToDelete.length);
+            let errorDelete = false;
+            // Batch accepts max 100 products per request
+            for (let i = 0; i < productsToDelete.length; i += 100) {
+                const batch = productsToDelete.slice(i, i + 100).map(p => p.id);
+                await shop.post("products/batch", { delete: batch }).then(async () => {
+                    // Success
+                    // delete batch from productsToDelete array and resave file (in case the process crashes, we can continue later)
+                    productsToDelete.splice(0, batch.length);
+                    fs.writeFileSync(productsToDeletePath, JSON.stringify(productsToDelete));
+                    deleteBar.update(i + batch.length);
+                }).catch((error) => {
+                    console.error('Failed to delete product batch in WooCommerce!');
+                    console.error(error);
+                    errorDelete = true;
+                });
+            }
 
-            await shop.post('products/batch', {
-                delete: productsToDelete.map(p => p.id)
-            }).then(() => {
+            if (!errorDelete) {
                 fs.unlink(productsToDeletePath, (err) => {
-                    console.log('Product deletion completed.');
                     if (err) console.error(`Error deleting file ${productsToDeletePath}:`, err);
                 });
-            })
-                .catch((error) => {
-                    console.error(`Error deleting products: `, error.response.data);
-                });;
-        } else console.log(`No products to delete.`);
+            } else {
+                console.log('Product deletion encountered errors. Please check the logs and re-run syncWoo() to continue.');
+            }
+        } else deleteBar.setTotal(0);
 
         if (productsQtyToUpdate.length > 0) {
-            console.log(`Starting quantity update for ${productsQtyToUpdate.length} products in WooCommerce [${shop.url}]...`);
-
             // Get fresh products from DB to make sure we have the latest quantity
             const freshProducts = await Product.find({ _id: { $in: productsQtyToUpdate.map(p => p._id) } });
 
-            await WooUpdateQuantityProducts(freshProducts, shop);
+            //TODO: Handle errors and partial updates
+            await WooUpdateQuantityProducts(freshProducts, shop, updateBar);
 
-            console.log(`Quantity update completed.`);
             fs.unlink(productsQtyToUpdatePath, (err) => {
                 if (err) console.error(`Error deleting file ${productsQtyToUpdatePath}:`, err);
             });
-        } else console.log(`No products to update.`);
+        } else updateBar.setTotal(0);
 
         if (productsToCreate.length > 0) {
-            console.log(`Creating ${productsToCreate.length} products in WooCommerce...`);
+            createBar.setTotal(productsToCreate.length);
 
             // Get fresh products from DB to make sure we have the latest data
             const freshProducts = await Product.find({ _id: { $in: productsToCreate.map(p => p._id) } });
@@ -122,18 +138,29 @@ async function syncWoo() {
 
                 await WooCreateProduct(product, shop);
 
+                createBar.increment();
                 // Remove product from array and resave file (in case the process crashes, we can continue later)
                 productsToCreate.shift();
                 fs.writeFileSync(productsToCreatePath, JSON.stringify(productsToCreate));
             }
-            console.log(`Products creation completed.`);
             fs.unlink(productsToCreatePath, (err) => {
                 if (err) console.error(`Error deleting file ${productsToCreatePath}:`, err);
             });
-        } else console.log(`No products to create.`);
+        } else createBar.setTotal(0);
+
+        return true;
     }
 
-    console.log("-----------------------------------------");
+    for (let shop of WooCommerce_Shops) {
+        const deleteBar = multibar.create(1, 0, { shop: shop.url, action: 'delete' });
+        const createBar = multibar.create(1, 0, { shop: shop.url, action: 'create' });
+        const updateBar = multibar.create(1, 0, { shop: shop.url, action: 'update' });
+        tasks.push(syncShopProducts(shop, WooCommerce_Shops.indexOf(shop), deleteBar, updateBar, createBar));
+    }
+
+    await Promise.all(tasks);
+
+    multibar.stop();
     console.log("WooCommerce sync completed.")
 }
 
@@ -142,24 +169,37 @@ async function compareAllProducts() {
 
     // Run createSnapshots() before running this function!
 
-    for (let i = 0; i < WooCommerce_Shops.length; i++) {
-        const shop = WooCommerce_Shops[i];
-        console.log(`----------------------------------------`);
-        console.log(`Comparing products for shop: ${shop.url}`);
+    const tasks = [];
+
+    const multibar = new MultiBar({
+        clearOnComplete: false,
+        hideCursor: true,
+        format: ' {bar} | {percentage}% | {value}/{total} {productType} Products | Shop: {shop}'
+    }, Presets.shades_classic);
+
+    let i = 0;
+    for (let shop of WooCommerce_Shops) {
         const wooProducts = JSON.parse(fs.readFileSync(`server/woocommerce/dev/results/shop_${i}_wooProducts.json`));
         const dbProducts = JSON.parse(fs.readFileSync(`server/woocommerce/dev/results/dbProducts.json`));
+        const barWoo = multibar.create(wooProducts.length, 0, { shop: shop.url, productType: 'Woo' });
+        const barDb = multibar.create(dbProducts.length, 0, { shop: shop.url, productType: 'DB' });
+        tasks.push(compareProducts(wooProducts, dbProducts, shop, barWoo, barDb, i));
+        i++;
+    }
 
-        console.log('Woo Products total: ' + wooProducts.length);
-        console.log('DB Products total: ' + dbProducts.length)
 
+
+    async function compareProducts(wooProducts, dbProducts, shop, barWoo, barDb, i) {
         if (!wooProducts || !dbProducts) {
             console.log(`No products found for shop: ${shop.url}. Make sure to run createSnapshots() first.`);
-            continue;
+            return true;
         }
 
         const productsToDelete = [];
         const productsToCreate = [];
         const productsQtyToUpdate = [];
+
+        const queue = new PQueue({ concurrency: 3 });
 
         for (let wooProduct of wooProducts) {
             // Check if any product exists in WooCommerce but not in DB. If it does, delete it from WooCommerce.
@@ -174,19 +214,26 @@ async function compareAllProducts() {
             // Check if quantity doesnt match
             if (wooProduct.type === 'simple' && dbProduct.quantity !== wooProduct.stock_quantity) {
                 productsQtyToUpdate.push(dbProduct);
+                barWoo.increment();
             } else if (wooProduct.type === 'variable') {
                 // Check each variation quantity
-                // TODO TEST BELOW CODE
-                const req = await shop.get(`products/${wooId}/variations`, { per_page: 100 });
-                const variations = req.data;
-
-                if (variations.some(v => v.stock_quantity !== dbProduct.sizes.find(s => s.woocommerce.find(w => w.woo_url === shop.url)?.id.toString() === v.id.toString())?.quantity)) {
-                    productsQtyToUpdate.push(dbProduct);
-                }
+                queue.add(async () => {
+                    const req = await shop.get(`products/${wooId}/variations`, { per_page: 100 });
+                    const variations = req.data;
+                    variations.forEach(async (variation) => {
+                        if (variation.stock_quantity !== dbProduct.sizes.find(s => s.woocommerce.find(w => w.woo_url === shop.url)?.id.toString() === variation.id.toString())?.quantity) {
+                            productsQtyToUpdate.push(dbProduct);
+                        }
+                    });
+                    barWoo.increment();
+                });
             }
         }
 
+        await queue.onIdle();
+
         for (let dbProduct of dbProducts) {
+            barDb.increment();
             // Check if any product exists in DB but not in WooCommerce. If it does, create it in WooCommerce.
             const wooId = dbProduct.woocommerce.find(w => w.woo_url === shop.url)?.id.toString();
             const wooProduct = wooProducts.find(p => p.id.toString() === wooId);
@@ -202,9 +249,8 @@ async function compareAllProducts() {
         const productsToCreatePath = `server/woocommerce/dev/results/productsToCreate_shop_${i}.json`;
 
         if (productsToDelete.length > 0) {
-            console.log(`Found ${productsToDelete.length} products in WooCommerce that do not exist in DB.`);
+            console.log(`Found ${productsToDelete.length} products in WooCommerce ${shop.url} that do not exist in DB. Will be deleted.`);
             fs.writeFileSync(productsToDeletePath, JSON.stringify(productsToDelete));
-            //TODO
         } else {
             console.log(`No products to delete in WooCommerce for shop: ${shop.url}`);
             if (fs.existsSync(productsToDeletePath)) {
@@ -213,10 +259,8 @@ async function compareAllProducts() {
         }
 
         if (productsQtyToUpdate.length > 0) {
-            console.log(`Found ${productsQtyToUpdate.length} products in WooCommerce that need updating.`);
+            console.log(`Found ${productsQtyToUpdate.length} products in WooCommerce ${shop.url} that need updating. Will be updated.`);
             fs.writeFileSync(productsQtyToUpdatePath, JSON.stringify(productsQtyToUpdate));
-            //TODO
-            // Instead of using the products in the array, grab them from the DB directly, because the qty may be different
         } else {
             console.log(`No products to update in WooCommerce for shop: ${shop.url}`);
             if (fs.existsSync(productsQtyToUpdatePath)) {
@@ -225,9 +269,8 @@ async function compareAllProducts() {
         }
 
         if (productsToCreate.length > 0) {
-            console.log(`Found ${productsToCreate.length} products in DB that do not exist in WooCommerce.`);
+            console.log(`Found ${productsToCreate.length} products in DB that do not exist in WooCommerce ${shop.url}. Will be created.`);
             fs.writeFileSync(productsToCreatePath, JSON.stringify(productsToCreate));
-            //TODO
         } else {
             console.log(`No products to create in WooCommerce for shop: ${shop.url}`);
             if (fs.existsSync(productsToCreatePath)) {
@@ -235,11 +278,15 @@ async function compareAllProducts() {
             }
         }
 
-        console.log(`----------------------------------------`);
+        return true;
     }
+
+    await Promise.all(tasks);
+    multibar.stop();
 
     console.log('All comparisons completed. Review the logs for details.');
 }
+
 async function createSnapshots() {
     // This functions saves all products from all WooCommerce shops and the DB to local files for later comparison.
     // Use this function before running any bulk comparisons
@@ -255,14 +302,23 @@ async function createSnapshots() {
     fs.writeFileSync(`server/woocommerce/dev/results/dbProducts.json`, JSON.stringify(dbProducts));
     console.log(`Saved ${dbProducts.length} products from DB to dbProducts.json`);
 
-    let i = 0;
-    for (let shop of WooCommerce_Shops) {
+    const multibar = new MultiBar({
+        clearOnComplete: false,
+        hideCursor: true,
+        format: ' {bar} | {percentage}% | {value}/{total} Products | Shop: {shop}'
+    }, Presets.shades_classic);
+
+    const tasks = [];
+
+    async function getProductsFromShop(shop, i, bar) {
         let done = false;
         let offset = 0;
         const wooProducts = [];
 
-        console.log(`Starting WooCommerce products batch get for shop: ${shop.url}`);
         while (done == false) {
+            if (offset + 100 > dbProducts.length)
+                bar.setTotal(offset + 100);
+
             const req = await shop.get("products", { per_page: 100, offset, orderby: 'id' });
             const products = req.data;
             if (products.length < 100) done = true;
@@ -270,21 +326,33 @@ async function createSnapshots() {
             wooProducts.push(...products);
 
             offset += 100;
-            console.log(`Got ${wooProducts.length} products from WooCommerce. Attempting to get more...`);
+            bar.update(wooProducts.length);
         }
+
+        bar.setTotal(wooProducts.length);
+        bar.update(wooProducts.length);
 
         // Save to file
         fs.writeFileSync(`server/woocommerce/dev/results/shop_${i}_wooProducts.json`, JSON.stringify(wooProducts));
-        console.log(`Saved ${wooProducts.length} products from WooCommerce to wooProducts.json`);
+        return true;
+    }
 
+    let i = 0;
+    for (let shop of WooCommerce_Shops) {
+        const bar = multibar.create(dbProducts.length, 0, { shop: shop.url });
+        tasks.push(getProductsFromShop(shop, i, bar));
         i++;
     }
 
+    await Promise.all(tasks);
+    multibar.stop();
     console.log('Snapshots created. You can now run compareAllProducts() to compare the data.');
 }
+
 async function fullySyncProducts() {
     // This function creates a snapshot of all products in the DB and WooCommerce, compares them and performs the necessary actions to fully sync them.
-    await createSnapshots();
-    await compareAllProducts();
+    // await createSnapshots();
+    // await compareAllProducts();
     await syncWoo();
 }
+fullySyncProducts();
